requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.44.0
django-debug-toolbar==3.2.2
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.4
idna==3.3
Markdown==3.3.4
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
sqlparse==0.4.2
urllib3==1.26.7
